@@ -1,14 +1,14 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.0"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 25
-        versionCode 43
-        versionName "4.0.3"
+        targetSdkVersion 28
+        versionCode 52
+        versionName "5.0.2"
     }
     buildTypes {
         release {
@@ -16,13 +16,19 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    lintOptions {
+        abortOnError false
+    }
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile 'com.android.support:support-v4:25.0.0'
-    compile 'com.soundcloud.android.crop:lib_crop:1.0.0'
-    compile 'com.darsh.multipleimageselect:multipleimageselect:1.0.4'
-    compile 'me.shaohui.advancedluban:library:1.3.2'
+    implementation 'androidx.annotation:annotation:1.1.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'me.shaohui.advancedluban:library:1.3.2'
+    implementation project(':lib_crop')//https://github.com/Shuaigoal/android-crop
+    implementation project(':lib_multipleimageselect')//https://github.com/Shuaigoal/MultipleImageSelect
+    implementation "androidx.exifinterface:exifinterface:1.0.0"
 }
 //apply from: "bintrayUpload.gradle"
