@@ -2,7 +2,8 @@ package com.jph.takephoto.app;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
+
 import android.util.Log;
 
 import com.jph.takephoto.R;
