@@ -7,7 +7,8 @@ import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
+
 import android.widget.Toast;
 
 import com.darsh.multipleimageselect.helpers.Constants;
