@@ -1,10 +1,11 @@
 package com.jph.simple;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentTransaction;
 import android.view.View;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentTransaction;
 
 
 /**
@@ -27,7 +28,7 @@ import android.view.View;
  * GitHub:https://github.com/crazycodeboy
  * Eamil:crazycodeboy@gmail.com
  */
-public class SimpleFragmentActivity extends FragmentActivity{
+public class SimpleFragmentActivity extends FragmentActivity {
     SimpleFragment fragment;
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
