@@ -2,7 +2,6 @@ package com.jph.takephoto.app;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.util.Log;
 
 import com.jph.takephoto.R;
@@ -12,6 +11,8 @@ import com.jph.takephoto.model.TResult;
 import com.jph.takephoto.permission.InvokeListener;
 import com.jph.takephoto.permission.PermissionManager;
 import com.jph.takephoto.permission.TakePhotoInvocationHandler;
+
+import androidx.fragment.app.Fragment;
 
 /**
  * 继承这个类来让Fragment获取拍照的能力<br>
